Changed gamemode name
I think people would get mad if they see my name in their server - so I
changed it. I had originally made it 'Tyguy's GunGame' because Lifepunch
had THEIR gungame gamemode name as "GunGame Lifepunch", and because
GunGame was in the name of the gamemode - my gamemode got tagged in it
too, which means that the Lifepunch gamemode and my gungame gamemode
were in the same place. And, as you can see, I changed the gamemode name
to Gun-Game, it looks better to me and better to you (hopefully)